fix(matriceal): use explicit int32_t casts in std::clamp for ARM build
ARM cross-compiler deduces conflicting types (long int vs int) when
mixing int32_t with literal 0/127. Cast literals to int32_t.

+++ src/deluge/model/sequencing/matriceal_engine.cpp
@@ -81,7 +81,7 @@ MatricealNote MatricealEngine::step(const MusicalKey& key) {
 	}
 
 	// Clamp to MIDI range
-	noteCode = std::clamp(noteCode, 0, 127);
+	noteCode = std::clamp(noteCode, (int32_t)0, (int32_t)127);
 
 	// Check probability — values 0-100 control per-step chance of note firing
 	if (probability.length > 0) {
